% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arc.R
\name{opac.arc}
\alias{opac.arc}
\title{Draw arc with opacity}
\usage{
opac.arc(x = 0.5, y = 0.5, r = 1, fill = "red", alpha.fill = 0.5,
  stroke.size = NULL, stroke.col = "blue", alpha.stroke = 0.5,
  start.angle = 0, end.angle = 90)
}
