% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{opac.ellipse}
\alias{opac.ellipse}
\title{Draw ellipse}
\usage{
opac.ellipse(x = 0.5, y = 0.5, width = 1, height = 1, fill = "red",
  alpha.fill = 0.5, stroke.size = 1, stroke.col = "blue",
  alpha.stroke = 0.5, rotate = 0)
}
