% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.R
\name{opac.pie}
\alias{opac.pie}
\title{Draw pie with opacity}
\usage{
opac.pie(x = 0.5, y = 0.5, r = 1, slices = c(10, 12, 4, 16, 8),
  labels = c("US", "UK", "Australia", "Germany", "France"), fill = NULL,
  alpha.fill = 0.5, stroke.size = 0.2, stroke.col = NULL,
  alpha.stroke = 1)
}
