% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{opac.point}
\alias{opac.point}
\title{Draw a point with opacity}
\usage{
opac.point(x, y = NULL, r = 1, fill = "skyblue", stroke.col = "black",
  stroke.size = 0.1, alpha.fill = 1, alpha.stroke = 1)
}
\arguments{
\item{x}{x coordinate}

\item{y}{y coordinate}

\item{r}{radius or size of the point}

\item{fill}{color for fill area}

\item{stroke.col}{color for stroke area}

\item{alpha.fill}{fill alpha}

\item{alpha.stroke}{stroke alpha}

\item{stroke.width}{width of stroke area}
}
\description{
Point with overlap between fill and stroke
}
\examples{
set.seed(1007)
demo.opac.point()
#----
# or
#----
plot.new()
opac.point(x=0.3, y=0.1, r = 0.5)
opac.point(x=0.3, y=0.9, r = 0.5, alpha.fill = 0.5, alpha.stroke = 0.5)
#------
opac.point(x=0.2, y=0.4, r = 0.5)
opac.point(x=0.2, y=0.45, r = 0.5, alpha.fill = 0.5, alpha.stroke = 0.5)
#------
opac.point(x=0.4, y=0.45, r = 0.5, alpha.fill = 0.5, alpha.stroke = 0.5)
opac.point(x=0.4, y=0.4, r = 0.5)
#------
opac.point(x=0.7, y=0.1, r = 0.5, alpha.fill = 0.5, alpha.stroke = 1)
opac.point(x=0.7, y=0.9, r = 0.5, alpha.fill = 1, alpha.stroke = 0.5, fill = 'orange3')
#---
opac.point(x=0.6, y=0.45, r = 0.5, alpha.fill = 1, alpha.stroke = 0.5, fill = 'orange3')
opac.point(x=0.6, y=0.4, r = 0.5, alpha.fill = 0.5, alpha.stroke = 1)
#---
opac.point(x=0.8, y=0.4, r = 0.5, alpha.fill = 0.5, alpha.stroke = 1)
opac.point(x=0.8, y=0.45, r = 0.5, alpha.fill = 1, alpha.stroke = 0.5, fill = 'orange3')
}
