% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rect.R
\name{opac.rect}
\alias{opac.rect}
\title{Draw rectangle with opacity}
\usage{
opac.rect(x = 0.5, y = 0.5, width = 3, height = 3, fill = "red",
  alpha.fill = 0.5, stroke.size = 1, stroke.col = "blue",
  alpha.stroke = 0.5, rotate = 0)
}
