% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle.R
\name{opac.triangle}
\alias{opac.triangle}
\title{Draw triangle with opacity}
\usage{
opac.triangle(x = 0.5, y = 0.5, width = 3, height = 3, fill = "red",
  alpha.fill = 0.5, stroke.size = 1, stroke.col = "blue",
  alpha.stroke = 0.5, rotate = 0)
}
